+++ cypress/support/api/index.js
@@ -1,2 +1,3 @@
 require('./auth'); // I have NO idea why but these need to be require not import
 require('./agreements');
+require('./licenses');